Separate test suites by php version
