+++ phpunit.xml.dist
@@ -15,7 +15,9 @@
 
     <testsuites>
         <testsuite name="Yii Injector tests">
-            <directory>./tests/</directory>
+            <directory>./tests/Common</directory>
+            <directory phpVersion="8" phpVersionOperator="&lt;">./tests/Php7</directory>
+            <directory phpVersion="8" phpVersionOperator=">=">./tests/Php8</directory>
         </testsuite>
     </testsuites>
