+++ tests/Common/MissingRequiredArgumentExceptionTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Injector\Tests;
+namespace Yiisoft\Injector\Tests\Common;
 
 class MissingRequiredArgumentExceptionTest extends ArgumentExceptionTest
 {
