+++ tests/Common/Support/EngineVAZ2101.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Injector\Tests\Support;
+namespace Yiisoft\Injector\Tests\Common\Support;
 
 class EngineVAZ2101 extends LightEngine
 {
